pyproject.toml
[tool.poetry]
name = "metagenerator"
version = "0.1.0"
description = ""
authors = ["Edmund Widl <edmund.widl@ait.ac.at>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.2"
jinja2 = "^3.1.6"
click = "^8.1.8"
schema = "^0.7.7"

[tool.poetry.scripts]
generator-helper = "metagenerator.cli.generator:main"
metagenerator-helper = "metagenerator.cli.metagenerator:main"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
